requirements.txt

# database
psycopg2-binary

# gdrive
pydrive
httplib2==0.15.0

# scrapping and parser
bs4
lxml

#Downloader
pyaiodl


requests

pyrogram
# https://github.com/pyrogram/pyrogram/archive/asyncio.zip
TgCrypto


